+++ site/src/editor/utils.rs
@@ -561,17 +561,14 @@ fn set_code_html(id: &str, code: &str) {
                                 .signature
                                 .map(|sig| sig.to_string())
                                 .unwrap_or_default();
-                            let class = if let Some(sig) = docs.signature {
-                                let margs = text.chars().rev().take_while(|c| *c == '!').count();
-                                binding_class(text, sig, margs, docs.constant)
-                            } else {
-                                ""
-                            };
+                            let class = binding_class(text, docs);
                             if let Some(comment) = &docs.comment {
                                 if !title.is_empty() {
                                     title.push(' ');
                                 }
                                 title.push_str(comment);
+                            } else if docs.is_module {
+                                title.push_str("module");
                             }
                             format!(
                                 r#"<span 
@@ -1019,6 +1016,7 @@ pub fn progressive_strings(input: &str) -> Vec<String> {
                 lines.push(vec![binding.span().as_str(&inputs, |s| s.into())])
             }
             Item::TestScope(items) => lines.push(vec![items.span.as_str(&inputs, |s| s.into())]),
+            Item::Import(import) => lines.push(vec![import.span().as_str(&inputs, |s| s.into())]),
         }
     }
     let mut strings = Vec::new();